+++ packages/auth/src/Domain/Token/ValetTokenData.ts
@@ -3,7 +3,10 @@ import { Uuid } from '@standardnotes/common'
 export type ValetTokenData = {
   userUuid: Uuid,
   permittedOperation: 'read' | 'write' | 'delete',
-  permittedResources: Array<string>,
+  permittedResources: Array<{
+    remoteIdentifier: string,
+    unencryptedFileSize?: number,
+  }>,
   uploadBytesUsed: number,
   uploadBytesLimit: number,
 }