+++ packages/features/src/Domain/Feature/Features.ts
@@ -414,12 +414,13 @@ function experimentalFeatures(): FeatureDescription[] {
   const markdownVisual: EditorFeatureDescription = FillEditorComponentDefaults({
     name: 'Markdown Visual (Beta)',
     identifier: FeatureIdentifier.MarkdownVisualEditor,
-    version: '1.0.1',
+    version: '1.0.4',
     note_type: NoteType.Markdown,
     file_type: 'md',
     permission_name: PermissionName.MarkdownVisualEditor,
     spellcheckControl: true,
-    description: 'A lightweight WYSIWYG markdown editor, derivated from Milkdown editor.',
+    description:
+      'A WYSIWYG-style Markdown editor that renders Markdown in preview-mode while you type without displaying any syntax.',
     git_repo_url: 'https://github.com/standardnotes/markdown-visual',
     marketing_url: 'https://github.com/standardnotes/markdown-visual',
     static_files: ['build'],