+++ packages/filepicker/src/types.ts
@@ -6,5 +6,5 @@ export type OnChunkCallback = (
 
 export type FileSelectionResponse = {
   name: string
-  ext: string
+  mimeType: string
 }